A small refactor here and there
